pyproject.toml
[project]
name = "transformer-jax"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dm-haiku>=0.0.13",
    "jax>=0.5.1",
]
